mtd: Fix broken dual parallel mutex logic
The dual parallel implementation wrongly tries to recursively call into
the same spi_nor_write function after it has taken the nor->lock mutex.
This results in waiting for the same mutex which is already held,
leading to a deadlock.

If no writes have been done, the read-path by itself is OK. After doing
a write on a dual-parallel flash, with the write starting from an odd
offset, things will go bad and the mutex will be deadlocked. After this
reads will fail as well.

The kernel prints look something like the following when reproducing the issue
and waiting for a few minutes:

  INFO: task dd:1225 blocked for more than 120 seconds.
        Not tainted 6.6.100-manual-g6372cdcf893a #27
  "echo 0 > /proc/sys/kernel/hung_task_timeout_secs" disables this message.
  task:dd              state:D stack:0     pid:1225  ppid:1214   flags:0x00000008
  Call trace:
   __switch_to+0xb8/0xdc
   __schedule+0x380/0xaf4
   schedule+0x54/0xe4
   schedule_preempt_disabled+0x14/0x24
   __mutex_lock+0x2d4/0x3f4
   mutex_lock_nested+0x20/0x28
   spi_nor_prep_and_lock_pe.part.0+0x38/0x11c
   spi_nor_write+0x9c/0x4b8
   spi_nor_write+0x114/0x4b8
   mtd_write_oob_std+0x70/0x78
   mtd_write_oob+0x68/0xac
   mtd_write+0x64/0xa0
   mtdchar_write+0x11c/0x27c
   vfs_write+0xdc/0x318
   ksys_write+0x6c/0x100
   __arm64_sys_write+0x18/0x20
   invoke_syscall.constprop.0+0x4c/0xe0
   do_el0_svc+0x3c/0xb8
   el0_svc+0x48/0x1a4
   el0t_64_sync_handler+0x118/0x124
   el0t_64_sync+0x14c/0x150
  INFO: task lmc_server:1231 blocked for more than 120 seconds.
        Not tainted 6.6.100-manual-g6372cdcf893a #27
  "echo 0 > /proc/sys/kernel/hung_task_timeout_secs" disables this message.
  task:lmc_server      state:D stack:0     pid:1231  ppid:913    flags:0x00000000
  Call trace:
   __switch_to+0xb8/0xdc
   __schedule+0x380/0xaf4
   schedule+0x54/0xe4
   schedule_preempt_disabled+0x14/0x24
   __mutex_lock+0x2d4/0x3f4
   mutex_lock_nested+0x20/0x28
   spi_nor_read+0x1ec/0x63c
   mtd_read_oob_std+0x74/0x7c
   mtd_read_oob+0x88/0x144
   mtd_read+0x64/0xa0
   mtdchar_read+0x228/0x2c0
   vfs_read+0xe4/0x2cc
   ksys_read+0x6c/0x100
   __arm64_sys_read+0x18/0x20
   invoke_syscall.constprop.0+0x4c/0xe0
   do_el0_svc+0x3c/0xb8
   el0_svc+0x48/0x1a4
   el0t_64_sync_handler+0x118/0x124
   el0t_64_sync+0x14c/0x150
  INFO: lockdep is turned off.

As a fix, move the mutex-taking after the dual-parallel logic point.
This is also in sync with how it was done in v6.1. This way the mutex
won't be re-entrantly taken in spi_nor_write.

While the read-path does not have such logic, also move the mutex-taking
after the dual-parallel special stuff to be aligned with the v6.1 logic.

Fixes: b696d5b74562 ("mtd: spi-nor: Add parallel memories support in spi-nor")
Signed-off-by: Jani Nurminen <jani.nurminen@windriver.com>
[Liyin: Fix the incorrect comments in the source code and the above commit
message. It should be odd 'offset', not 'length'.]
Signed-off-by: Liyin Zhang <liyin.zhang.cn@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>