rdinit: fix boot warning caused by warning supression
The commit: 3b2f001e6a236 [boot: only emit rdinit warning on initramfs
boot] was causing boot failures on configurations that use initramfs
as it incorrectly changed the boot flow while silencing the warning
for machines not using an initramfs.

With this, we get the warning silenced and working initramfs boots.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>