AIR-13630 neutron: Reset NPU on mailbox send failure
commit  f9b4954952b6db0db55096125bb9614ea223c54e from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When sending data to the mailbox fails, the hardware may be in an error
state, which could cause subsequent operations to malfunction or hang.

To mitigate this issue, call the neutron_hw_reset() function in the
error handling path to restore the device to a normal state.

Signed-off-by: Jiwei.Fu <jiwei.fu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>