spi: spi-zynqmp-gqspi: Fix random padding issue when the data length to be written exceeds 256 bytes
When the data length to be written exceeds 256 bytes, the
spi-zynqmp-gqspi driver issues GENFIFO entries using a combination of
exponent and immediate modes, with the immediate transfers aligned to
4 bytes. Accordingly, when filling the TX FIFO, the data is also
padded to achieve 4-byte alignment.
However, when filling the last incomplete 4-byte chunk of xqspi->txbuf,
the local variable intermediate is not initialized, causing the padding
bytes to contain random values.
Initialize intermediate to 0xFFFFFFFF to avoid this issue.
In addition, after the final memcpy, count should be assigned first
before clearing xqspi->bytes_to_transfer.

Signed-off-by: Liyin Zhang <liyin.zhang.cn@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>