Revert "spi: spi-zynqmp-gqspi: Fix memory corruption when writing odd-length data in dual parallel mode"
This reverts commit e74766cbdbf05ace7a9555e695f45b2af2457874.
In this commit, the op->data.buf.out buffer is strictly allocated in
the MTD layer by the mtdchar_write() function based on the write
length provided for each call, and then passed down to the QSPI driver.
The QSPI driver attempted to modify this buffer and write to it in
zynqmp_qspi_exec_op() to achieve 4-byte alignment padding.
As a result, this causes slab-out-of-bounds and memory corruption
issues.
Instead, the padding bytes (0xFF) should be handled when writing to
the TX FIFO.

Signed-off-by: Liyin Zhang <liyin.zhang.cn@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>