TI: HACK: arm64: boot: dts: ti: k3-am62l-wakeup: Configure AM62L pinctrl for wakeup interrupts
commit 3e80672f3884d1752caebb8ae0e8eacb69473406 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add interrupts to the pinctrl node in order to allow I/O daisy chain
wakeups. Based on the patch for AM625.

Marked as HACK because physically none of the wakeup source events are
connected to GIC. Eventually we may want to set up a chained interrupt
handler in the ti-msgmgr for pinctrl-single to use.

Signed-off-by: Kendall Willis <k-willis@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>