TI: arm64: boot: dts: ti: Create AM62L LPM wakeup sources overlay
commit 0917002ac111a2fd10b4ff64be0900d9390406f7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Make an overlay for AM62L so that new wakeup sources can be added for
low power modes.

This specifically enables WKUP GPIO with the WKUP UART RXD signal as the
wakeup. This wakeup source is separated from the k3-am62l3-evm.dts because
WKUP UART wakeup is enabled in that device tree, so using the WKUP UART
RXD signal as WKUP GPIO would not be possible.

Signed-off-by: Kendall Willis <k-willis@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>