PENDING: pmdomain: Add braces to include constraint_ns in if condition
commit 7a1070b67da60a869b49d194f729473b64c9020a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Due to missing braces around the if condition, constraint_ns was not set
correctly. This caused the following build warnings:

drivers/pmdomain/governor.c: In function ‘cpu_system_power_down_ok’:
drivers/pmdomain/governor.c:449:17: warning: this ‘if’ clause does not guard... [-Wmisleading-indentation]
  449 |                 if (min_dev_latency_ns < constraint_ns)
      |                 ^~
drivers/pmdomain/governor.c:452:25: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the ‘if’
  452 |                         constraint_ns = min_dev_latency_ns;
      |

Add the braces to include constraint_ns as part of the if condition, and
also update the comment above to be more descriptive of this behaviour.

Fixes: f5e72a936ec9 ("PENDING: pmdomain: consider per-device QoS when checking system power down state")
Signed-off-by: Dhruva Gole <d-gole@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>