PENDING: arm64: dts: ti: k3-am62l3-evm: Move DPI pinctrl to HDMI bridge node
commit 95b10e92a9ef287da60333051dfcc89ea8fca798 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Move the DPI pinctrl configuration from the DSS node to the sii9022 HDMI
bridge node.

This change ensures the pinctrl configuration is associated with the
actual consumer of the DPI pins, which is the HDMI bridge device
rather than the DSS controller. This also helps avoid potential pin
ownership conflicts and ensures the pins are configured when the bridge
driver probes.

Signed-off-by: Swamil Jain <s-jain1@ti.com>
Reviewed-by: Devarsh Thakkar <devarsht@ti.com>
Tested-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>