octeontx2-af: map management port always to first PF
commit 66d235c16d5bdc0b5f525d6d1830a0dba3674334 from
git@git.assembla.com:cavium/WindRiver.linux.git

The user can enable or disable any MAC block or a few ports of the
block. The management port's interface name varies depending on the
setup of the user if its not mapped to the first pf.

By changing the management port mapping to always point to the first PF,
this patch resolves the problem.

Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Change-Id: I23d04eb9b0b84c527045dd7060b7a2bcbb2588c2
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>