crypto: inside-secure: Backward compatible to legacy module param
commit ec5229d5735daed043e55e20488b0e16fb9338a1 from
git@git.assembla.com:cavium/WindRiver.linux.git

In earlier SDK kernels (v4.14, v5.4), we replaced the module parameter
'max_rings=<n>" with 'rings=<x,y,z>' in order to support the different
rings allocation for each engine. It caused significant changes to the
mainline driver but may be not necessary for MUSDK user.

This patch added the module param 'rings' back to be backward compatible
to old SDKs while it still keeps 'max_ring' to be aligned with the
mainline driver.

Signed-off-by: Wilson Ding <dingwei@marvell.com>
Change-Id: If5ec7cdec8c546fbd17d16fb5da482c156c86912
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/148580
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sanghoon Lee <salee@marvell.com>
Signed-off-by: kmahammadraf <kmahammadraf@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/168027
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>