LF-16818 net: enetc: reset PIR and CIR if they are not equal when initializing TX ring
commit  bf5bcc0111a535e80a6dba83f0b716908cf92ef6 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Currently, the hardware PIR and CIR are not initialized by the software
when the TX BD is re-initialized. The driver just reads hardware PIR and
CIR and then initializes the software PIR and CIR. Therefore, if the TX
BD ring is re-initialized when it still has some unsent frames, there
will be a problem that the TX BD ring cannot work properly after the
re-initialization. Because the hardware PIR and CIR are not equal and
the unsent frames have been freed. The BDs between CIR and PIR are
invalid, which will cause a hardware malfunction.

For ENETC v4, the solution is very simple, just reset the hardware PIR
and CIR if they are not equal when initializing the TX BD ring. But for
ENETC v1 (LS1028A), the solution is not feasible due to the hardware
difference. And currently we have not a solution for ENETC v1, so this
patch is only applicable to ENETC v4.

Fixes: 99100d0d9922 ("net: enetc: add preliminary support for i.MX95 ENETC PF")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>