LF-17115 media: i2c: ox03c10: Reset VS exposure before starting the stream
commit  45b3500614903a6bc07c8bd8520dad5de75d2bc5 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

It appears that resetting the VS to 4 after streaming is stopped, to
avoid a bad image next time we start the streaming, does not work
reliably all the time. Especially when FSYNC is used and pipelines are
stopped and restarted repeatedly.

However, resetting VS just before starting streaming works fine. There
is no clear explanation on why this happens as there is no clear
explanation from OV on why starting streaming with a VS setting higher
than 4 messes up the image. But, for now, this is the only solution to
the problem.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>