scsi: ufs: amd-versal2: Fix PHY initialization in HCE enable notify
commit 990c956ecf5245b1f4a5998475f09d016e8ad268 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

Move the PHY initialization from PRE_CHANGE to POST_CHANGE in the
ufs_versal2_hce_enable_notify() callback. This ensures that the PHY
is initialized after the host controller enable sequence is complete,
rather than before it starts.

The PHY initialization requires the UFS host controller to be in a
stable enabled state to properly configure the MPHY registers. Moving
this to POST_CHANGE aligns with the expected initialization order and
prevents potential timing issues during controller startup.

Signed-off-by: Ajay Neeli <ajay.neeli@amd.com>
Link: https://lore.kernel.org/all/20251224053950.54213-1-ajay.neeli@amd.com
State: waiting
Message-ID: <20251224061246.116170-1-ajay.neeli@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>